still fixing

+++ install.sh
@@ -69,8 +69,10 @@ if [ ! -d "~/.local/share/fonts/" ]; then
   mkdir -p ~/.local/share/fonts/
 fi
 
-curl -fLo "DejaVu Sans Mono Nerd Font Complete.ttf" https://github.com/ryanoasis/nerd-fonts/raw/master/patched-fonts/DejaVuSansMono/Regular/complete/DejaVu%20Sans%20Mono%20Nerd%20Font%20Complete.ttf
-mv DejaVu\ Sans\ Mono\ Nerd\ Font\ Complete.ttf $HOME/.local/share/fonts/
+if [ ! -f "~/.local/share/fonts/DejaVu Sans Mono Nerd Font Complete.ttf" ]; then
+  curl -fLo "DejaVu Sans Mono Nerd Font Complete.ttf" https://github.com/ryanoasis/nerd-fonts/raw/master/patched-fonts/DejaVuSansMono/Regular/complete/DejaVu%20Sans%20Mono%20Nerd%20Font%20Complete.ttf
+  mv DejaVu\ Sans\ Mono\ Nerd\ Font\ Complete.ttf $HOME/.local/share/fonts/
+fi
 
 printf "${GREEN}DONE!${NORMAL}\n"
